+++ frontend/src/models/groupContext.ts
@@ -1,7 +1,0 @@
-export type groupContext = {
-  gc_id: number;
-  group_id: number;
-  gc_seqn: number;
-  gc_name: string;
-  gc_text: string;
-};