+++ frontend/src/pages/elements/RAGDocuments/groupLLMs.ts
@@ -1,6 +1,0 @@
-export type GroupLLMs = {
-  gllms_id: number;
-  group_id: number;
-  gllms_name: string;
-  gllms_status: "success" | "warning" | "danger";
-};